refactor: move editCurseModal to modal folder
