+++ frontend/src/components/adminPage/editCourse/modal/editCourseModal.module.scss
